% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-annotate.R
\name{cn_setup}
\alias{cn_setup}
\title{Download and set up DepMap CN}
\usage{
cn_setup(
  overwrite = TRUE,
  data_dir = system.file("extdata", package = "gimap")
)
}
\arguments{
\item{overwrite}{Should the files be redownloaded?}

\item{data_dir}{What directory should this be saved to? Default is with package
files}
}
\description{
This function sets up the tpm data from DepMap is called
by the `gimap_annotate()` function if the cn_annotate = TRUE
}
