% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encrypt_creds_path}
\alias{encrypt_creds_path}
\title{Default creds path}
\usage{
encrypt_creds_path(app_name)
}
\arguments{
\item{app_name}{What app set up are you looking for? Supported apps are 'google' 'calendly' and 'github'}
}
\description{
Default creds path
}
