% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_figshare}
\alias{get_figshare}
\title{Handler function for GET requests from Figshare}
\usage{
get_figshare(
  file_name = NA,
  item = "19700056",
  output_dir = tempdir(),
  return_list = FALSE
)
}
\arguments{
\item{file_name}{Which item are we downloading?}

\item{item}{What is the item we are retrieving?}

\item{output_dir}{Where should the file be saved?}

\item{return_list}{Should the list of files be returned instead of the file}
}
\value{
Downloads necessary annotation files from Figshare and reads them
in as data frames.
}
\description{
Handler function for GET requests from Figshare
}
\examples{
\donttest{

get_figshare(
  return_list = TRUE,
  output_dir = tempdir()
)

get_figshare(
  file_name = "Achilles_common_essentials.csv",
  output_dir = tempdir()
)
}
}
