% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-setup_data.R
\name{gimap_object}
\alias{gimap_object}
\title{Make an empty gimap dataset object}
\usage{
gimap_object()
}
\value{
an empty `gimap_dataset` which is a named list which will
be filled by various `gimap` functions.
}
\description{
This function makes an empty gimap data object
}
