% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-annotate.R
\name{normalize_depmap_metadata}
\alias{normalize_depmap_metadata}
\title{Normalize DepMap metadata for downstream joins}
\usage{
normalize_depmap_metadata(depmap_metadata)
}
\description{
Normalize DepMap metadata for downstream joins
}
\keyword{internal}
