% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc.R
\name{plot_theme}
\alias{plot_theme}
\title{Standardized plot theme}
\usage{
plot_theme()
}
\value{
A ggplot2 theme that can be used on the plots.
}
\description{
this is a ggplot2 standardized plot theme for this package
}
