% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc.R
\name{qc_cdf}
\alias{qc_cdf}
\title{Create a CDF for the pgRNA normalized counts}
\usage{
qc_cdf(gimap_dataset, wide_ar = 0.75)
}
\arguments{
\item{gimap_dataset}{The special gimap_dataset from the `setup_data` function
which contains the transformed data}

\item{wide_ar}{aspect ratio, default is 0.75}
}
\value{
counts_cdf a ggplot
}
\description{
This function uses pivot_longer to rearrange the data for plotting
 and then plots a CDF of the normalized counts
}
\examples{
\dontrun{

gimap_dataset <- get_example_data("gimap")
qc_cdf(gimap_dataset)
}
}
