% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc.R
\name{qc_sample_hist}
\alias{qc_sample_hist}
\title{Create a histogram for the pgRNA log2 CPMs, faceted by sample}
\usage{
qc_sample_hist(gimap_dataset, wide_ar = 0.75)
}
\arguments{
\item{gimap_dataset}{The special gimap_dataset from the `setup_data` function
which contains the transformed data}

\item{wide_ar}{aspect ratio, default is 0.75}
}
\value{
sample_cpm_histogram a ggplot
}
\description{
This function uses pivot_longer to rearrange the data for plotting and then
plots sample specific histograms of the pgRNA cpm's
}
\examples{
\dontrun{
gimap_dataset <- get_example_data("gimap")
qc_sample_hist(gimap_dataset)
}
}
