% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_example_data_dir}
\alias{resolve_example_data_dir}
\title{Default directory for [get_example_data()] and related helpers}
\usage{
resolve_example_data_dir(data_dir)
}
\description{
Default directory for [get_example_data()] and related helpers
}
\keyword{internal}
