% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_example_timepoint_data}
\alias{save_example_timepoint_data}
\title{Set up example count data}
\usage{
save_example_timepoint_data(data_dir = NULL)
}
\arguments{
\item{data_dir}{Optional data directory override.}
}
\value{
Returns the file path to folder where the example data is stored
}
\description{
Set up example count data
}
