% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_example_treatment_data}
\alias{save_example_treatment_data}
\title{Set up example count data}
\usage{
save_example_treatment_data()
}
\value{
Returns the file path to folder where the example data is stored
}
\description{
Set up example count data
}
