% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-annotate.R
\name{supported_cell_lines}
\alias{supported_cell_lines}
\title{List the supported cell lines}
\usage{
supported_cell_lines()
}
\value{
A list of the cell line names that are available in DepMap for use
for annotation in this package.
}
\description{
This function downloads the metadata for DepMap and lists which
cell lines are supported.
}
\examples{

cell_lines <- supported_cell_lines()

}
