% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-annotate.R
\name{ctrl_genes}
\alias{ctrl_genes}
\title{Download and set up control genes}
\usage{
ctrl_genes(
  overwrite = TRUE,
  data_dir = system.file("extdata", package = "gimap")
)
}
\arguments{
\item{overwrite}{Should the file be redownloaded and reset up?}

\item{data_dir}{What directory should this be saved to? Default is with package
files}
}
\description{
This function sets up the control genes file from DepMap is
called by the `gimap_annotate()`
}
