% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_example_data}
\alias{delete_example_data}
\title{Refresh the example data files by redownloading them}
\usage{
delete_example_data()
}
\value{
options for example data are are set to NULL.
}
\description{
This function will set example data file options to NULL so files
will be re-downloaded
}
\examples{

delete_example_data()

}
