% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_data_folder}
\alias{example_data_folder}
\title{Get file path to an default credentials RDS}
\usage{
example_data_folder()
}
\value{
Returns the file path to folder where the example data is stored
}
\description{
Get file path to an default credentials RDS
}
